+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `DescriptionLocalizations` and `NameLocalizations` structs
+
+### Changed
+- `desc` and `desc_localizations` are now mutually exclusive
+- `desc_localizations` and `name_localizations` must return the
+  `DescriptionLocalizations` and `NameLocalizations` structs,
+  respectively.
 
 ## [0.15.2] - 2023-06-23
 ### Added